rename: 메일 서비스의 클래스명 변경
