+++ src/main/java/co/fineants/api/infra/mail/EmailSender.java
@@ -1,14 +1,14 @@
-package co.fineants.api.infra.mail.service;
+package co.fineants.api.infra.mail;
 
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.stereotype.Service;
+import org.springframework.stereotype.Component;
 
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-@Service
-public class MailService {
+@Component
+public class EmailSender {
 	private final JavaMailSender mailSender;
 
 	public void sendEmail(String to, String subject, String body) {
